Add a socket timeout for the IRC connection
Kaet has issues where the TCP connection to IRC will drop and the bot
itself won't notice. We'll fix this by adding a socket timeout to the
IRC connection.
Go's treatment of socket timeouts is apparently like a watchdog: you
set a timeout to be a time, not a duration, and then continue updating
it as you successfully receive or send data.
This code assumes a twitch IRC PING or other input comes in at least
once every 5 minutes. If that's not the case, we'll see more frequent
bot restarts and want to bump the timeout value in feedWatchdog.
I set up my own environment, and when I only set the Deadline on initial
connection, it times out after 5 minutes, but with the set on receive,
it stays connected. If I shut down my wifi, the process exits after the
timeout has been exceeded.

Fixes #1